Module Smc_inference.Interruptible

The Interruptible module exposes an implementation of Sequential Monte-Carlo that gives back the hand to the user after each resampling.

type 'a resumption =

| Finished of ('a * Log_space.t) list

| Suspended of ?resampling:resampling_strategy -> int -> 'a resumption

val run : ?resampling:resampling_strategy -> int -> 'a t -> RNG.t -> 'a resumption

run npart model rng_state runs sequential Monte-Carlo with npart initial particles. resampling defaults to systematic_resampling.

raises Invalid_population
if the total mass of the population becomes equal to Log_space.zero