Module Smc_inference.Non_interruptible

The Non_interruptible module exposes an implementation of Sequential Monte-Carlo that terminates when all particles terminate. Only use when the probabilistic program terminates almost surely!

val run : ?resampling:resampling_strategy -> int -> 'a t -> RNG.t -> ('a * Log_space.t) list

run ?resampling npart model rng_state runs sequential Monte-Carlo with npart initial particles. resampling defaults to systematic_resampling.

raises Invalid_population
if the total mass of the population becomes equal to Log_space.zero