Module type Smc_inference.Resampling_types_S

The SMC implementation is parametric in the type of values produced by the particles at each yield and in the type of state of resampling.

type particle_output

The type of values given to yield.

type resampling_state

The type of state on which resampling steps operate.